inserita java-doc

+++ src/main/java/com/aremi/springclustersimservice/util/TaskResult.java
@@ -3,9 +3,13 @@ package com.aremi.springclustersimservice.util;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+/**
+ * Classe di utilità che contiene il risultato di un task
+ */
+
 @Getter
 @RequiredArgsConstructor
 public class TaskResult {
     private final boolean isSuccess;
-    private final int executionTime;
+    private final long executionTime;
 }
